Reorganizando pacotes (detalhes na descricao)
- Ao inves de ter um pacote diferente para cada classe, juntei todas num pacote chamado "classe". Sao poucas classes, portanto, nao deve ficar confuso jogar todas no mesmo lugar.

- Transformei o enum Pagamentos em um HashMap, assim simplifica pedir ao usuario qual a forma de pagamento. Se continuasse como enum, eu teria que pedir um numero e fazer mais um switch/case para procurar o pagamento correto. Essa funcionalidade poderia ser feita com um simples ArrayList, porem prefiri usar HashMap, pois ele me da mais flexibilidade se eu precisar mudar como eu adquiro a forma de pagamento
