--- Provas/P2/pizzaria/src/br/maua/enumerates/Pagamento.java
@@ -1,5 +1,0 @@
-package br.maua.enumerates;
-
-public enum Pagamento {
-    DINHEIRO, DEBITO, CREDITO, VALE_ALIMENTACAO, VALE_REFEICAO
-}
